independent configcenter source
